Refactored the structure of the test extensions
